requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==22.10.0
celery==5.4.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==4.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.4.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pathspec==0.10.1
pika==1.3.1
platformdirs==2.5.3
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.5.0
pymongo==3.12.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.13
